Trait corustine::Coroutine

pub trait Coroutine<T>: Debug {
    fn init(&mut self, scheduler: &Scheduler<T>);

    fn resume(&mut self, message: T) -> Pause<T>;
}

This trait defines a coroutine in order to be used with this crate.

Required Methods

fn init(&mut self, scheduler: &Scheduler<T>)

Initializes self, so it can, as an example, grab all necessary ids.

fn resume(&mut self, message: T) -> Pause<T>

Resumes/starts this coroutine.